chore(news):move News component to NewsList component
