+++ src/js/components/NewsList.jsx
@@ -1,12 +1,26 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import '../../scss/main.scss';
+import News from './News';
 
 /**
  * @param {any} props
  * @return {element} list of all articles
  */
-const NewsList = props => <ul className="flex-container">{props.news}</ul>;
+const NewsList = (props) => {
+  const news = props.news;
+  const NewsComponents = news.map((item, i) => {
+    let title;
+    if (item.title.length > 60) {
+      title = `${item.title.slice(0, 56)}...`;
+    } else {
+      title = item.title;
+    }
+    const key = i;
+    return <News key={key} title={title} href={item.url} src={item.urlToImage} />;
+  });
+  return <ul className="flex-container">{NewsComponents}</ul>;
+};
 
 /*eslint-disable*/
 NewsList.propTypes = {
